+++ R/h_survival_duration_subgroups.R
@@ -56,12 +56,10 @@ NULL
 #'
 #' @export
 h_survtime_df <- function(tte, is_event, arm) {
-  assertthat::assert_that(
-    is.numeric(tte),
-    is.logical(is_event),
-    is.factor(arm),
-    is_equal_length(tte, is_event, arm)
-  )
+  checkmate::assert_numeric(tte)
+  checkmate::assert_flag(is_event)
+  assert_valid_factor(arm)
+  assert_equal_length(tte, is_event, arm)
 
   df_tte <- data.frame(
     tte = tte,
@@ -189,13 +187,11 @@ h_survtime_subgroups_df <- function(variables,
 #'
 #' @export
 h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control_coxph()) {
-  assertthat::assert_that(
-    is.numeric(tte),
-    is.logical(is_event),
-    is.factor(arm),
-    assertthat::are_equal(nlevels(arm), 2),
-    is_equal_length(tte, is_event, arm)
-  )
+  checkmate::assert_numeric(tte)
+  checkmate::assert_flag(is_event)
+  assert_valid_factor(arm)
+  checkmate::assert_set_equal(nlevels(arm), 2)
+  assert_equal_length(tte, is_event, arm)
 
   df_tte <- data.frame(tte = tte, is_event = is_event)
   strata_vars <- NULL