+++ R/prop_diff.R
@@ -24,18 +24,15 @@ check_diff_prop_ci <- function(rsp,
                                strata = NULL,
                                conf_level,
                                correct = NULL) {
-  assertthat::assert_that(
-    is.logical(rsp),
-    !anyNA(c(rsp, grp)),
-    is_equal_length(rsp, grp),
-    nlevels(grp) == 2,
-    conf_level >= 0,
-    conf_level <= 1
-  )
+  checkmate::assert(!anyNA(c(rsp, grp)))
+  checkmate::assert_logical(rsp)
+  checkmate::assert_set_equal(nlevels(grp), 2)
+  checkmate::assert_number(conf_level, lower = 0, upper = 1)
+  assert_equal_length(rsp, grp)
 
-  if (!is.null(correct)) assertthat::assert_that(assertthat::is.flag(correct))
+  if (!is.null(correct))checkmate::assert_flag(correct)
 
-  if (!is.null(strata)) assertthat::assert_that(is_equal_length(rsp, strata))
+  if (!is.null(strata)) assert_equal_length(rsp, strata)
 
   invisible()
 }